Create a base tensor train class
